backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
firebase-admin==6.5.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
prophet==1.1.5
torch==2.4.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.5.0
httpx==0.27.0
Pillow==10.4.0
python-dateutil==3.0.0
